mods to get compile to work on samd mkr1000

+++ src/sfeTkArdSPI.cpp
@@ -138,7 +138,7 @@ int32_t sfeTkArdSPI::writeRegisterRegion(uint8_t devReg, const uint8_t *data, si
     digitalWrite(cs(), LOW);
     _spiPort->transfer(devReg);
 
-    for (int i = 0; i < length; i++)
+    for (size_t i = 0; i < length; i++)
         _spiPort->transfer(*data++);
 
     // End communication
